backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
Pillow
django-cors-headers
drf-nested-routers